// packages/backend/src/auth/enums/role.enum.ts
// ==============================================================================
// Enumeração de Funções (Roles)
// Autor: Deivid Lucas
// Versão: 1.1
// Descrição: Define as funções de utilizador disponíveis no sistema.
//            Adicionada a função USER como padrão para novos registos.
// ==============================================================================

export enum Role {
  ADMIN = 'ADMIN',
  ANALISTA = 'ANALISTA',
  AUDITOR = 'AUDITOR',
  CEO = 'CEO',
  CFO = 'CFO',
  CLIENTE = 'CLIENTE',
  COMPRADOR = 'COMPRADOR',
  COORDENADOR = 'COORDENADOR',
  COO = 'COO',
  CTO = 'CTO',
  DESENVOLVEDOR = 'DESENVOLVEDOR',
  DESENVOLVIMENTO = 'DESENVOLVIMENTO',
  DESIGNER = 'DESIGNER',
  DIRETOR = 'DIRETOR',
  ESTAGIARIO = 'ESTAGIARIO',
  ESTRATEGIA = 'ESTRATEGIA',
  FATURAMENTO = 'FATURAMENTO',
  FINANCEIRO = 'FINANCEIRO',
  FREELANCER = 'FREELANCER',
  GERENTE = 'GERENTE',
  GESTOR_ESTOQUE = 'GESTOR_ESTOQUE',
  INOVACAO = 'INOVACAO',
  JURIDICO = 'JURIDICO',
  LIDER = 'LIDER',
  LOGISTICA = 'LOGISTICA',
  MANUTENCAO = 'MANUTENCAO',
  MARKETING = 'MARKETING',
  MOTORISTA = 'MOTORISTA',
  OPERADOR = 'OPERADOR',
  PESQUISA = 'PESQUISA',
  PLANEJAMENTO = 'PLANEJAMENTO',
  PRODUCAO = 'PRODUCAO',
  QUALIDADE = 'QUALIDADE',
  RELACIONAMENTO = 'RELACIONAMENTO',
  RH = 'RH',
  SUPERADMIN = 'SUPER_ADMIN',
  SUPERVISOR = 'SUPERVISOR',
  SUPORTE = 'SUPORTE',
  TECNICO = 'TECNICO',
  TESTADOR = 'TESTADOR',
  TREINAMENTO = 'TREINAMENTO',
  USER = 'USER',
  VENDEDOR = 'VENDEDOR',
}